Remove old test file and move tests to new one.
